move TheatrePrinter.cs into AppUI/TheatrePrinters folder
